+++ cypress/integration/demo.ts
@@ -13,15 +13,15 @@ describe('Rule editor demo', () => {
     it('should successfully upload a JSON file', () => {
       cy.get('select#questionnaire-select').select('Upload your own questionnaire');
 
-      cy.get('#link-id').type('/39156-5');
       cy.get('#file-upload').attachFile('bmi.json');
 
       cy.get('lhc-rule-editor h1').contains('Rule Editor');
-      cy.get('#final-expression').should('have.value', '(%a/(%b.power(2))).round(1)');
+
+      cy.get('#final-expression').should('not.exist');
 
       // Updating the linkId should update the Rule Editor instantly
-      cy.get('#link-id').type('a');
-      cy.get('#final-expression').should('not.exist');
+      cy.get('select#link-id').select('/39156-5');
+      cy.get('#final-expression').should('have.value', '(%a/(%b.power(2))).round(1)');
     });
   });
 });