+++ e2e/src/app.po.ts
@@ -41,21 +41,6 @@ export class AppPage {
     return element.all(by.css('.remove-variable')).get(index).click();
   }
 
-  async addQuestionVariable(question, unit): Promise<void> {
-    await this.addVariable();
-    const index = await this.getNumberOfVariables() - 1;
-
-    // Set as question
-    await element(by.id('variable-type-' + index)).click();
-    await element(by.id(`row-${index}`)).element(by.cssContainingText('option', 'Question')).click();
-    await element(by.css(`#row-${index} .question-select`)).click();
-    await element(by.id(`row-${index}`)).element(by.cssContainingText('option', question)).click();
-    await element(by.css(`#row-${index} .unit-select`)).click();
-    await element(by.id(`row-${index}`)).element(by.cssContainingText('option', unit)).click();
-
-    return Promise.resolve();
-  }
-
   async checkAddVariable(): Promise<boolean> {
     const variablesBefore = await this.getNumberOfVariables();
     await this.addVariable();
@@ -70,10 +55,6 @@ export class AppPage {
     const variablesAfter = await this.getNumberOfVariables();
 
     return Promise.resolve(variablesBefore - variablesAfter === 1);
-  }
-
-  async setFinalExpression(expression): Promise<void> {
-    return element(by.id('final-expression')).sendKeys(expression);
   }
 
   async setSimpleExpression(expression): Promise<void> {